+++ src/screens/twinClasses/index.ts
@@ -1,0 +1,1 @@
+export { TwinClasses } from "./twin-classes";
